move ListToSPInfo to /api/lists.go
